move ScalingField into scapy.fields
